Decompose app/main.py into focused middleware/lifecycle modules
Extract request logging/redaction into app/middleware/request_logging.py, exception handlers into app/exception_handlers.py, and DB startup logic into app/lifecycle.py. Replace sys.exit(1) in the app factory with raise RuntimeError. Update tests to reference the moved helpers and new logger name. main.py shrinks from 766 to ~340 lines.
